Virtual Machine Overhaul: 64-bit Precision and Scalability
- Transition VM numerics from 32-bit float to 64-bit double architecture.
- Optimize object initialization using DefaultValuesArray and Array.Copy.
- Move built-in variable indices to ObjectType for O(1) access.
- Increase all engine limits by several orders of magnitude (MaxStack, MaxGlobals, etc).
- Unroll interpreter loop (8x) and use raw pointer bytecode access.
- Ensure DM parity with epsilon equality for double comparisons.
- Update test suite to support 8-byte numeric constants.

+++ Developers/Compiler/Optimizer/AnnotatedBytecode.cs
@@ -196,9 +196,9 @@ internal sealed class AnnotatedBytecodeInteger(int value, Location location) : I
     }
 }
 
-internal sealed class AnnotatedBytecodeFloat(float value, Location location) : IAnnotatedBytecode {
+internal sealed class AnnotatedBytecodeFloat(double value, Location location) : IAnnotatedBytecode {
     public Location Location = location;
-    public float Value = value;
+    public double Value = value;
 
     public void AddArg(DMCompiler compiler, IAnnotatedBytecode arg) {
         compiler.ForcedError(Location, "Cannot add args to a float");
